# client/openapi.json
{"openapi":"3.0.1","info":{"title":"DriMain API","description":"Unified API for Web + Mobile clients","version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/zgloszenia/{id}":{"get":{"tags":["zgloszenie-rest-controller"],"operationId":"get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ZgloszenieDTO"}}}}}},"put":{"tags":["zgloszenie-rest-controller"],"operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ZgloszenieUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ZgloszenieDTO"}}}}}},"delete":{"tags":["zgloszenie-rest-controller"],"operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"No Content"}}}},"/api/raporty/{id}":{"get":{"tags":["raport-rest-controller"],"operationId":"get_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RaportDTO"}}}}}},"put":{"tags":["raport-rest-controller"],"operationId":"update_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RaportUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RaportDTO"}}}}}},"delete":{"tags":["raport-rest-controller"],"operationId":"delete_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"No Content"}}}},"/api/czesci/{id}":{"get":{"tags":["part-rest-controller"],"operationId":"get_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PartDTO"}}}}}},"put":{"tags":["part-rest-controller"],"operationId":"update_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PartUpdateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PartDTO"}}}}}},"delete":{"tags":["part-rest-controller"],"operationId":"delete_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"No Content"}}}},"/api/zgloszenia":{"get":{"tags":["zgloszenie-rest-controller"],"operationId":"list","parameters":[{"name":"status","in":"query","required":false,"schema":{"type":"string"}},{"name":"typ","in":"query","required":false,"schema":{"type":"string"}},{"name":"q","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ZgloszenieDTO"}}}}}}},"post":{"tags":["zgloszenie-rest-controller"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ZgloszenieCreateRequest"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ZgloszenieDTO"}}}}}}},"/api/raporty":{"get":{"tags":["raport-rest-controller"],"operationId":"list_1","parameters":[{"name":"status","in":"query","required":false,"schema":{"type":"string"}},{"name":"maszynaId","in":"query","required":false,"schema":{"type":"integer","format":"int64"}},{"name":"from","in":"query","required":false,"schema":{"type":"string","format":"date"}},{"name":"to","in":"query","required":false,"schema":{"type":"string","format":"date"}},{"name":"q","in":"query","required":false,"schema":{"type":"string"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":25}},{"name":"sort","in":"query","required":false,"schema":{"type":"string","default":"dataNaprawy,desc"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageRaportDTO"}}}}}},"post":{"tags":["raport-rest-controller"],"operationId":"create_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RaportCreateRequest"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RaportDTO"}}}}}}},"/api/czesci":{"get":{"tags":["part-rest-controller"],"operationId":"list_2","parameters":[{"name":"kat","in":"query","required":false,"schema":{"type":"string"}},{"name":"q","in":"query","required":false,"schema":{"type":"string"}},{"name":"belowMin","in":"query","required":false,"schema":{"type":"boolean"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PartDTO"}}}}}}},"post":{"tags":["part-rest-controller"],"operationId":"create_2","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PartCreateRequest"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PartDTO"}}}}}}},"/api/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/czesci/{id}/ilosc":{"patch":{"tags":["part-rest-controller"],"operationId":"adjust","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PartQuantityPatch"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PartDTO"}}}}}}},"/api/meta/statusy/zgloszenia":{"get":{"tags":["meta-controller"],"operationId":"zgloszenieStatuses","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"type":"string"}}}}}}}},"/api/meta/statusy/raporty":{"get":{"tags":["meta-controller"],"operationId":"raportStatuses","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"type":"string"}}}}}}}},"/api/auth/me":{"get":{"tags":["auth-controller"],"operationId":"me","parameters":[{"name":"Authorization","in":"header","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}}},"components":{"schemas":{"ZgloszenieUpdateRequest":{"type":"object","properties":{"typ":{"type":"string"},"imie":{"type":"string"},"nazwisko":{"type":"string"},"tytul":{"type":"string"},"status":{"type":"string"},"opis":{"type":"string"},"dzialId":{"type":"integer","format":"int64"},"autorId":{"type":"integer","format":"int64"},"photoBase64":{"type":"string"}}},"ZgloszenieDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"typ":{"type":"string"},"imie":{"type":"string"},"nazwisko":{"type":"string"},"tytul":{"type":"string"},"status":{"type":"string","enum":["OPEN","IN_PROGRESS","ON_HOLD","DONE","REJECTED"]},"opis":{"type":"string"},"dataGodzina":{"type":"string","format":"date-time"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"dzialId":{"type":"integer","format":"int64"},"dzialNazwa":{"type":"string"},"autorId":{"type":"integer","format":"int64"},"autorUsername":{"type":"string"},"hasPhoto":{"type":"boolean"}}},"PartUsageDTO":{"type":"object","properties":{"partId":{"type":"integer","format":"int64"},"ilosc":{"type":"integer","format":"int32"}}},"RaportUpdateRequest":{"type":"object","properties":{"typNaprawy":{"type":"string"},"opis":{"type":"string"},"status":{"type":"string"},"dataNaprawy":{"type":"string","format":"date"},"czasOd":{"type":"string"},"czasDo":{"type":"string"},"maszynaId":{"type":"integer","format":"int64"},"osobaId":{"type":"integer","format":"int64"},"partUsages":{"type":"array","items":{"$ref":"#/components/schemas/PartUsageDTO"}}}},"RaportDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"maszyna":{"$ref":"#/components/schemas/SimpleMaszynaDTO"},"typNaprawy":{"type":"string"},"opis":{"type":"string"},"status":{"type":"string"},"dataNaprawy":{"type":"string","format":"date"},"czasOd":{"type":"string"},"czasDo":{"type":"string"},"osoba":{"$ref":"#/components/schemas/SimpleOsobaDTO"},"partUsages":{"type":"array","items":{"$ref":"#/components/schemas/PartUsageDTO"}}}},"SimpleMaszynaDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nazwa":{"type":"string"}}},"SimpleOsobaDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"imieNazwisko":{"type":"string"}}},"PartUpdateRequest":{"type":"object","properties":{"nazwa":{"type":"string"},"kod":{"type":"string"},"kategoria":{"type":"string"},"minIlosc":{"type":"integer","format":"int32"},"jednostka":{"type":"string"}}},"PartDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nazwa":{"type":"string"},"kod":{"type":"string"},"kategoria":{"type":"string"},"ilosc":{"type":"integer","format":"int32"},"minIlosc":{"type":"integer","format":"int32"},"jednostka":{"type":"string"}}},"ZgloszenieCreateRequest":{"type":"object","properties":{"typ":{"type":"string"},"imie":{"type":"string"},"nazwisko":{"type":"string"},"tytul":{"type":"string"},"status":{"type":"string"},"opis":{"type":"string"},"dataGodzina":{"type":"string","format":"date-time"},"dzialId":{"type":"integer","format":"int64"},"autorId":{"type":"integer","format":"int64"},"photoBase64":{"type":"string"}}},"RaportCreateRequest":{"type":"object","properties":{"maszynaId":{"type":"integer","format":"int64"},"typNaprawy":{"type":"string"},"opis":{"type":"string"},"osobaId":{"type":"integer","format":"int64"},"status":{"type":"string"},"dataNaprawy":{"type":"string","format":"date"},"czasOd":{"type":"string"},"czasDo":{"type":"string"},"partUsages":{"type":"array","items":{"$ref":"#/components/schemas/PartUsageDTO"}}}},"PartCreateRequest":{"type":"object","properties":{"nazwa":{"type":"string"},"kod":{"type":"string"},"kategoria":{"type":"string"},"ilosc":{"type":"integer","format":"int32"},"minIlosc":{"type":"integer","format":"int32"},"jednostka":{"type":"string"}}},"AuthRequest":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"PartQuantityPatch":{"type":"object","properties":{"delta":{"type":"integer","format":"int32"}}},"PageRaportDTO":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/RaportDTO"}},"number":{"type":"integer","format":"int32"},"sort":{"type":"array","items":{"$ref":"#/components/schemas/SortObject"}},"numberOfElements":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"offset":{"type":"integer","format":"int64"},"sort":{"type":"array","items":{"$ref":"#/components/schemas/SortObject"}},"paged":{"type":"boolean"},"unpaged":{"type":"boolean"}}},"SortObject":{"type":"object","properties":{"direction":{"type":"string"},"nullHandling":{"type":"string"},"ascending":{"type":"boolean"},"property":{"type":"string"},"ignoreCase":{"type":"boolean"}}}}}}